+++ Tests/SpotifyWebAPIMainTests/API Tests/SpotifyAPILibraryTests.swift
@@ -1,0 +1,254 @@
+import Foundation
+import XCTest
+import Combine
+@testable import SpotifyWebAPI
+import SpotifyAPITestUtilities
+import SpotifyExampleContent
+
+protocol SpotifyAPILibraryTests: SpotifyAPITests { }
+
+extension SpotifyAPILibraryTests where
+    AuthorizationManager: SpotifyScopeAuthorizationManager
+{
+    
+    func saveAlbums() {
+        
+        let fullAlbums = URIs.Albums.array(
+            .skiptracing, .housesOfTheHoly, .tiger, .illmatic
+        )
+        
+        let partialAlbunms = URIs.Albums.array(
+            .skiptracing, .housesOfTheHoly
+        )
+        
+        let expectation = XCTestExpectation(description: "testSaveAlbums")
+        
+        Self.spotify.removeSavedAlbumsForCurrentUser(fullAlbums)
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedAlbums(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .flatMap { savedAlbumsArray -> AnyPublisher<Void, Error> in
+                let allAlbums = savedAlbumsArray
+                    .flatMap(\.items).compactMap(\.item.uri)
+                for album in allAlbums {
+                    XCTAssertFalse(fullAlbums.contains(album))
+                }
+                return Self.spotify.saveAlbumsForCurrentUser(partialAlbunms)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap { () -> AnyPublisher<[Bool], Error> in
+                Self.spotify.currentUserSavedAlbumsContains(fullAlbums)
+            }
+            .XCTAssertNoFailure()
+            .flatMap { results -> AnyPublisher<Void, Error> in
+                XCTAssertEqual(results, [true, true, false, false])
+                return Self.spotify.removeSavedAlbumsForCurrentUser(fullAlbums)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedAlbums(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .sink(
+                receiveCompletion: { _ in expectation.fulfill() },
+                receiveValue: { savedAlbumsArray in
+                    let allAlbums = savedAlbumsArray
+                        .flatMap(\.items).compactMap(\.item.uri)
+                    for album in allAlbums {
+                        XCTAssertFalse(fullAlbums.contains(album))
+                    }
+                }
+            )
+            .store(in: &Self.cancellables)
+        
+        self.wait(for: [expectation], timeout: 300)
+            
+    }
+    
+    func saveTracks() {
+        
+        let fullTracks = URIs.Tracks.array(
+            .because, .blueBoy, .breathe, .faces
+        )
+        
+        let partialTracks = URIs.Tracks.array(
+            .because, .blueBoy
+        )
+        
+        let expectation = XCTestExpectation(description: "testSaveTracks")
+        
+        Self.spotify.removeSavedTracksForCurrentUser(fullTracks)
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedTracks(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .flatMap { savedTracksArray -> AnyPublisher<Void, Error> in
+                let allTracks = savedTracksArray
+                    .flatMap(\.items).compactMap(\.item.uri)
+                for track in allTracks {
+                    XCTAssertFalse(fullTracks.contains(track))
+                }
+                return Self.spotify.saveTracksForCurrentUser(partialTracks)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap { () -> AnyPublisher<[Bool], Error> in
+                Self.spotify.currentUserSavedTracksContains(fullTracks)
+            }
+            .XCTAssertNoFailure()
+            .flatMap { results -> AnyPublisher<Void, Error> in
+                XCTAssertEqual(results, [true, true, false, false])
+                return Self.spotify.removeSavedTracksForCurrentUser(fullTracks)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedTracks(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .sink(
+                receiveCompletion: { _ in expectation.fulfill() },
+                receiveValue: { savedTracksArray in
+                    let allTracks = savedTracksArray
+                        .flatMap(\.items).compactMap(\.item.uri)
+                    for track in allTracks {
+                        XCTAssertFalse(fullTracks.contains(track))
+                    }
+                }
+            )
+            .store(in: &Self.cancellables)
+        
+        self.wait(for: [expectation], timeout: 300)
+            
+    }
+    
+    func saveShows() {
+        
+        let fullShows = URIs.Shows.array(
+            .joeRogan, .samHarris, .scienceSalon, .seanCarroll
+        )
+        
+        let partialShows = URIs.Shows.array(
+            .joeRogan, .samHarris
+        )
+        
+        let expectation = XCTestExpectation(description: "testSaveTracks")
+        
+        Self.spotify.removeSavedShowsForCurrentUser(fullShows)
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedShows(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .flatMap { savedShowsArray -> AnyPublisher<Void, Error> in
+                let allShows = savedShowsArray
+                    .flatMap(\.items).map(\.item.uri)
+                for show in allShows {
+                    XCTAssertFalse(fullShows.contains(show))
+                }
+                return Self.spotify.saveShowsForCurrentUser(partialShows)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap { () -> AnyPublisher<[Bool], Error> in
+                Self.spotify.currentUserSavedShowsContains(fullShows)
+            }
+            .XCTAssertNoFailure()
+            .flatMap { results -> AnyPublisher<Void, Error> in
+                XCTAssertEqual(results, [true, true, false, false])
+                return Self.spotify.removeSavedShowsForCurrentUser(fullShows)
+            }
+            .XCTAssertNoFailure()
+            .delay(for: 1, scheduler: DispatchQueue.main)
+            .flatMap {
+                Self.spotify.currentUserSavedShows(
+                    limit: 50, offset: 0
+                )
+            }
+            .XCTAssertNoFailure()
+            .extendPages(Self.spotify)
+            .XCTAssertNoFailure()
+            .collect()
+            .sink(
+                receiveCompletion: { _ in expectation.fulfill() },
+                receiveValue: { savedShowsArray in
+                    let allShows = savedShowsArray
+                        .flatMap(\.items).map(\.item.uri)
+                    for show in allShows {
+                        XCTAssertFalse(fullShows.contains(show))
+                    }
+                }
+            )
+            .store(in: &Self.cancellables)
+        
+        self.wait(for: [expectation], timeout: 300)
+            
+    }
+    
+    
+}
+
+final class SpotifyAPIAuthorizationCodeFlowLibraryTests:
+    SpotifyAPIAuthorizationCodeFlowTests, SpotifyAPILibraryTests
+{
+    
+    static let allTests = [
+        ("testSaveAlbums", testSaveAlbums),
+        ("testSaveTracks", testSaveTracks),
+        ("testSaveShows", testSaveShows)
+    ]
+    
+    func testSaveAlbums() { saveAlbums() }
+    func testSaveTracks() { saveTracks() }
+    func testSaveShows() { saveShows() }
+
+}
+
+final class SpotifyAPIAuthorizationCodeFlowPKCELibraryTests:
+    SpotifyAPIAuthorizationCodeFlowPKCETests, SpotifyAPILibraryTests
+{
+    
+    static let allTests = [
+        ("testSaveAlbums", testSaveAlbums),
+        ("testSaveTracks", testSaveTracks),
+        ("testSaveShows", testSaveShows)
+    ]
+    
+    func testSaveAlbums() { saveAlbums() }
+    func testSaveTracks() { saveTracks() }
+    func testSaveShows() { saveShows() }
+
+}