+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.9.1] - 2020-10-27
+
+### Fixed
+
+Fixed bugs in which the following methods called the wrong endpoints:
+
+* `SpotifyAPI.currentUserSavedTracks(limit:offset:market:)`
+* `SpotifyAPI.currentUserSavedShowsContains(_:)`
+
+All of the methods in `SpotifyAPI` are now covered by unit tests!
+
 ## [0.9.0] - 2020-10-27
 
 ### Added